+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * All money types automatically format on blur:  values will be updated, if necessary, to always end in a period followed by two digits
 * Relaxed the int and float types/subtypes, and added strict versions of each.  The relaxed versions accept leading/trailing zeros and negative zero; the strict versions do not.
 * Added a logging callback function as a configurable option
-* Added unit tests
+* Added Selenium unit tests
 
 
 # Release 1.0.1 (April 2, 2016)